+++ src/handlers/metrics.rs
@@ -1,0 +1,29 @@
+//! Contains the `/metrics` endpoint filter.
+
+use crate::metrics::Metrics;
+use axum::body::HttpBody;
+use axum::routing::get;
+use axum::Router;
+
+pub trait MetricsRoutes {
+    /// Provides an API for Prometheus/OpenMetrics metrics.
+    ///
+    /// ```http
+    /// GET /metrics HTTP/1.1
+    /// ```
+    fn map_metrics_endpoint(self) -> Self;
+}
+
+impl<S, B> MetricsRoutes for Router<S, B>
+where
+    S: Clone + Send + Sync + 'static,
+    B: HttpBody + Send + 'static,
+{
+    fn map_metrics_endpoint(self) -> Self {
+        self.route("/metrics", get(render_metrics))
+    }
+}
+
+async fn render_metrics() -> String {
+    Metrics::get().encode()
+}