+++ src/handlers/mod.rs
@@ -1,0 +1,11 @@
+//! Contains warp filters.
+
+mod health;
+mod metrics;
+mod shutdown;
+mod yeet;
+
+pub use health::HealthRoutes;
+pub use metrics::MetricsRoutes;
+pub use shutdown::ShutdownRoutes;
+pub use yeet::YeetRoutes;