+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   `/livez`, as well as `/health` and `/healthz`.
 - Prometheus/OpenMetrics metrics is available at `/metrics`.
 - Shutdown from SIGINT (e.g. CTRL-C), SIGTERM and similar is now possible.
+- The `transfer_size_total` and `transfer_total` metrics now track the number of bytes sent through the service.
 
 ### Internal